refactor(test): import the runner's project list instead of parsing it
Review's point: the drift guard regex-parsed scripts/test.js for PROJECTS and
WATCH_DEFAULT, then needed its own cases for arrays across lines, for spreads
and for lists built by a call — tests of the test infrastructure's test
infrastructure. The reason given was that the runner cannot be imported, but
that is only true of the runner. testArgs.js already exists precisely because
it has no import-time side effects.

So the constants moved there and the guard imports both sides. What that
deleted: two regex readers, three cases pinning the regex's own behaviour, and
the throw that existed only because a regex can stop matching. Net 142 lines
out, 74 in, and a renamed displayName or a retyped constant now fails to
compile rather than failing to match.

It needed one more thing to work, and it is the same fix as version.ts:
jest.config.ts named-imported compilerOptions from tsconfig.json, which is a
bundler convenience rather than ESM, so importing the config from an esm-project
test failed with "does not provide an export named 'compilerOptions'". Default
import now. Jest still loads the config as CommonJS, where esModuleInterop
makes the two equivalent.

Two smaller points from the same review.

--selectProjects unit test/util/version.test.ts swallows the path into the
project list. Checked against jest rather than assumed: jest does the same —
`jest --listTests --selectProjects unit test/util/version.test.ts` lists the
whole unit project, filter and all. Stopping at the first positional would make
the runner do something jest does not, which is the worse surprise, so it is
pinned as deliberate with the = form noted as the way out.

And the coverage comment claimed a caller supplying --coverageDirectory "gets
one report". That was true of the single-process design and is not true now:
separate processes have nothing to merge, so a shared directory gets the
overwrite back. It is an escape hatch for redirecting output, not a combined
report, and there is no longer any such thing.

npm test: 109 suites, 1369 tests. build, tsc and eslint exit 0.
